refactor: HashMap에서 EnumMap으로 수정

+++ src/main/java/domain/Rank.java
@@ -28,7 +28,7 @@ public enum Rank {
         return matchCount;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
 }
